server: Refactor and  add routes for /events and /events/{id}
- /events lists the names, ids, and mimetypes of all events.
Created for MVP and will likely just be used for
debugging in the future
- /events/id lists the name, fileId, mimetype, and
textContent for a given file.
Created for MVP and will likely just be used for
debugging in the future. Also displays text results in
markdown.

💘💘💘💘💘💘💘💘💘💘
Did a refactoring of everything to take advantage of class
inheritance. Current vision is....

There will be multiple types of resources. Like 'events'
and 'albums'. But they wil all pull from Google Drive in
the same way
- Contract DriveService (interface) for shared functions
... for retreiving information from google drive.

- Abstract Class BaseDriveService
... which will be initialized with CWD (folderId). Since
all these 'DriveServices' are interacting in the same acct
and service (Google Drive) they should probably share this
logic. This class also includes a framework for interacting
with the drive itself.

- Class EventService
... implements the functions from the DriveService
interface like listAllResources and getResource. A future
e.g. AlbumService would look very similar, but with queries
tweaked.

🏁🏁🏁🏁🏁🏁🏁🏁🏁🏁